feat: add utility function to convert raw value on detail section

+++ core/src/main/java/com/bagusmerta/core/data/source/remote/MovieeResponse/MovieeDetailResponse.kt
@@ -69,7 +69,7 @@ data class MovieeDetailResponse(
     var tagline: String?,
 
     @Json(name = "vote_count")
-    var voteCount: Double?,
+    var voteCount: Int?,
 
     @Json(name = "original_language")
     var originalLanguage: String?,
